Clean up sqlite flags in 24Q3
This removes the sqlite trunk-stable flags from the build/release
directories in 24Q3.  These directories are used in 25Q2 and later but
not in 24Q3.  The flag values are removed to avoid confusion.

Tested by booting an image on cuttlefish and verifying that the
sqlite version ws 3.44.5.

Flag: RELEASE_PACKAGE_LIBSQLITE3
Bug: 430889718
Bug: 281064726
Test: presubmit
Change-Id: I5f7183f2222638beb6a8067dead7f935b4848a11